+++ Sources/StreamChat/APIClient/Endpoints/EndpointPath+OfflineRequest.swift
@@ -7,7 +7,7 @@ import Foundation
 extension EndpointPath {
     var shouldBeQueuedOffline: Bool {
         switch self {
-        case .sendMessage, .editMessage, .deleteMessage, .addReaction, .deleteReaction:
+        case .sendMessage, .editMessage, .deleteMessage, .pinMessage, .unpinMessage, .addReaction, .deleteReaction:
             return true
         case .createChannel, .connect, .sync, .users, .guest, .members, .search, .devices, .channels, .updateChannel,
              .deleteChannel, .channelUpdate, .muteChannel, .showChannel, .truncateChannel, .markChannelRead, .markChannelUnread,